Updated typos in the file
I fixed a type in code of the test function., which should not lead to different results. I started the windows two pixels too high, so te bottom 4 pixels where not used (this is only 1% of the window region)

Some code for evaluation is better suited for more classes and imbalance.